op5: import ramdisk from stock oos

+++ rootdir/etc/init.qcom.sh
@@ -1,5 +1,5 @@
 #!/system/bin/sh
-# Copyright (c) 2009-2017, The Linux Foundation. All rights reserved.
+# Copyright (c) 2009-2016, The Linux Foundation. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
@@ -25,6 +25,35 @@
 # OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
 # ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
+
+oemdump=`getprop persist.oem.dump`
+buildtype=`getprop ro.build.release_type`
+if [ "$oemdump" == "" ]; then
+    case "$buildtype" in
+        "release" | "cta")
+           setprop persist.oem.dump 0
+           ;;
+        *)
+           setprop persist.oem.dump 1
+           ;;
+    esac
+fi
+
+#check build variant for printk logging
+#current default minimum boot-time-default
+buildvariant=`getprop ro.build.type`
+case "$buildvariant" in
+    "userdebug" | "eng")
+        #set default loglevel to KERN_INFO
+        #Modify by david@bsp, 20161101 change console loglevel to 7
+        echo "7 6 1 7" > /proc/sys/kernel/printk
+        ;;
+    *)
+        #set default loglevel to KERN_WARNING
+        echo "4 4 1 4" > /proc/sys/kernel/printk
+        ;;
+esac
+
 
 target=`getprop ro.board.platform`
 if [ -f /sys/devices/soc0/soc_id ]; then
@@ -82,7 +111,7 @@ start_msm_irqbalance_8939()
 {
 	if [ -f /system/bin/msm_irqbalance ]; then
 		case "$platformid" in
-		    "239" | "293" | "294" | "295" | "304" | "313" | "338")
+		    "239" | "293" | "294" | "295" | "304" | "313")
 			start msm_irqbalance;;
 		esac
 	fi
@@ -167,6 +196,39 @@ case "$target" in
                 ;;
         esac
         start_charger_monitor
+        ;;
+    "sdm660")
+        if [ -f /sys/devices/soc0/soc_id ]; then
+            soc_id=`cat /sys/devices/soc0/soc_id`
+        else
+            soc_id=`cat /sys/devices/system/soc/soc0/id`
+        fi
+
+        if [ -f /sys/devices/soc0/hw_platform ]; then
+             hw_platform=`cat /sys/devices/soc0/hw_platform`
+        else
+             hw_platform=`cat /sys/devices/system/soc/soc0/hw_platform`
+        fi
+
+        case "$soc_id" in
+             "317" | "324" | "325" | "326" | "318" | "327" )
+                  case "$hw_platform" in
+                       "Surf")
+                                    setprop qemu.hw.mainkeys 0
+                                    ;;
+                       "MTP")
+                                    setprop qemu.hw.mainkeys 0
+                                    ;;
+                       "RCM")
+                                    setprop qemu.hw.mainkeys 0
+                                    ;;
+                       "QRD")
+                                    setprop qemu.hw.mainkeys 0
+                                    ;;
+                  esac
+                  ;;
+       esac
+        start_msm_irqbalance
         ;;
     "apq8084")
         platformvalue=`cat /sys/devices/soc0/hw_platform`
@@ -289,7 +351,7 @@ case "$target" in
              hw_platform=`cat /sys/devices/system/soc/soc0/hw_platform`
         fi
         case "$soc_id" in
-             "293" | "304" | "338" )
+             "293" | "304" )
                   case "$hw_platform" in
                        "Surf")
                                     setprop qemu.hw.mainkeys 0
@@ -304,7 +366,8 @@ case "$target" in
                   ;;
        esac
         ;;
-esac
+    esac
+
 
 #
 # Copy qcril.db if needed for RIL
@@ -315,40 +378,30 @@ echo 1 > /data/misc/radio/db_check_done
 #
 # Make modem config folder and copy firmware config to that folder for RIL
 #
-#ifdef VENDOR_EDIT
-# modify the source path to /system/etc/firmware/mbn_ota/ , hanqingpu, 20151119
-##if [ -f /data/misc/radio/ver_info.txt ]; then
-##    prev_version_info=`cat /data/misc/radio/ver_info.txt`
-##else
-##    prev_version_info=""
-##fi
+if [ -f /data/misc/radio/ver_info.txt ]; then
+    prev_version_info=`cat /data/misc/radio/ver_info.txt`
+else
+    prev_version_info=""
+fi
 
-##cur_version_info=`cat /firmware/verinfo/ver_info.txt`
-##if [ ! -f /firmware/verinfo/ver_info.txt -o "$prev_version_info" != "$cur_version_info" ]; then
-rm -rf /data/misc/radio/modem_config
-mkdir /data/misc/radio/modem_config
-chmod 770 /data/misc/radio/modem_config
-##    cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
-cp -r /system/etc/firmware/mbn_ota/* /data/misc/radio/modem_config
-chown -hR radio.radio /data/misc/radio/modem_config
-##    cp /firmware/verinfo/ver_info.txt /data/misc/radio/ver_info.txt
-##    chown radio.radio /data/misc/radio/ver_info.txt
-##fi
-##cp /firmware/image/modem_pr/mbn_ota.txt /data/misc/radio/modem_config
-##chown radio.radio /data/misc/radio/modem_config/mbn_ota.txt
+cur_version_info=`cat /firmware/verinfo/ver_info.txt`
+if [ ! -f /firmware/verinfo/ver_info.txt -o "$prev_version_info" != "$cur_version_info" ]; then
+    rm -rf /data/misc/radio/modem_config
+    mkdir /data/misc/radio/modem_config
+    chmod 770 /data/misc/radio/modem_config
+    cp -r /firmware/image/modem_pr/mcfg/configs/* /data/misc/radio/modem_config
+#ifdef VENDOR_EDIT
+# add for mbn_ota.txt , hanqingpu, 20161207
+    cp -r /system/etc/firmware/mbn_ota/mbn_ota.txt /data/misc/radio/modem_config/mbn_ota.txt
 #endif /*VENDOR_EDIT*/
+    chown -hR radio.radio /data/misc/radio/modem_config
+    cp /firmware/verinfo/ver_info.txt /data/misc/radio/ver_info.txt
+    chown radio.radio /data/misc/radio/ver_info.txt
+fi
+cp /firmware/image/modem_pr/mbn_ota.txt /data/misc/radio/modem_config
+chown radio.radio /data/misc/radio/modem_config/mbn_ota.txt
 echo 1 > /data/misc/radio/copy_complete
 
 #check build variant for printk logging
 #current default minimum boot-time-default
-buildvariant=`getprop ro.build.type`
-case "$buildvariant" in
-    "userdebug" | "eng")
-        #set default loglevel to KERN_INFO
-        echo "6 6 1 7" > /proc/sys/kernel/printk
-        ;;
-    *)
-        #set default loglevel to KERN_WARNING
-        echo "4 4 1 4" > /proc/sys/kernel/printk
-        ;;
-esac
+
